Modify: header token type from token to authorization bearer
