+++ routes/middleware/setAccessLevel.js
@@ -2,14 +2,15 @@ const mongoose = require("mongoose");
 const createError = require("http-errors");
 
 const User = require("../../models/User");
-const { verifyToken } = require("../utils/tokens");
+const { verifyToken, parseBearer } = require("../../helpers/tokens");
 const ACCESS_LEVELS = require("../../constants/accessLevels");
 const { NOT_FOUND } = require("../../constants/statusCodes");
 const { ERROR } = require("../../constants/messages");
 
 async function setAccessLevel(req, res, next) {
   const { creatorId } = req.params;
-  const { token: accessToken } = req.headers;
+  const { authorization } = req.headers;
+  const accessToken = parseBearer(authorization);
 
   try {
     if (!mongoose.Types.ObjectId.isValid(creatorId)) {
